Added concern and refactored the code
