+++ app/models/category.rb
@@ -1,11 +1,5 @@
 class Category < ActiveRecord::Base
+  include Image
   has_many  :products
   
-  validates :name, presence: true
-  has_attached_file :image, :styles => { :medium => "300x300>", :thumb => "100x100>" }, :default_url => "/images/:style/noimage.jpg"
-  validates_attachment_content_type :image, :content_type => /\Aimage\/.*\Z/
-
-  def image_delete
-    @image_delete ||= "0"
-  end
 end
