+++ app/views/categories/_form.html.erb
@@ -1,39 +1,39 @@
-  <table class="table">
-    <%= form_for(@category) do |f| %>
-      <tr class="row">
-        <th class="col-md-4">
-          <%= f.label :name %>
-        </th>
-        <td class="col-md-4">
-          <%= f.text_field :name, class: "col-md-12" %>
-        </td>
-      </tr>
-      <tr class="row">
-        <th class="col-md-4">
-          <%= f.label :description %>
-        </th>
-        <td class="col-md-4">
-          <%= f.text_area :description, class: "col-md-12" %>
-        </td>
-      </tr>
-      <tr class="row">
-        <th class="col-md-4">
-          <%= f.label :image %>
-        </th>
-        <td class="col-md-4">
-          <%= image_tag @category.image.url(:medium) %>
-          <%= f.file_field :image %>
-        </td>
-      </tr>
-      <% if @category.image.url != "/images/noimage.jpg" %>
-        <tr class="row">
-          <td></td>
-          <th class="col-md-3 col-md-offset-4"> <%= f.check_box :image_delete %> Delete image?</th>
-        </tr>
-      <% end %>
+<table class="table">
+  <%= form_for(@category) do |f| %>
+    <tr class="row">
+      <th class="col-md-4">
+        <%= f.label :name %>
+      </th>
+      <td class="col-md-4">
+        <%= f.text_field :name, class: "col-md-12" %>
+      </td>
+    </tr>
+    <tr class="row">
+      <th class="col-md-4">
+        <%= f.label :description %>
+      </th>
+      <td class="col-md-4">
+        <%= f.text_area :description, class: "col-md-12" %>
+      </td>
+    </tr>
+    <tr class="row">
+      <th class="col-md-4">
+        <%= f.label :image %>
+      </th>
+      <td class="col-md-4">
+        <%= image_tag @category.image.url(:medium) %>
+        <%= f.file_field :image %>
+      </td>
+    </tr>
+    <% if @category.image.url != "/images/noimage.jpg" %>
       <tr class="row">
         <td></td>
-        <td><%= f.submit "Save", class: "btn btn-primary" %></td>
+        <th class="col-md-3 col-md-offset-4"> <%= f.check_box :image_delete %> Delete image?</th>
       </tr>
     <% end %>
-  </table>
+    <tr class="row">
+      <td></td>
+      <td><%= f.submit "Save", class: "btn btn-primary" %></td>
+    </tr>
+  <% end %>
+</table>
