+++ app/views/categories/edit.html.erb
@@ -1,7 +1,5 @@
 <div class='container'>
   <h1>Edit Category</h1>
-
-  <!-- Render form here -->
   <%= link_to 'Back', categories_path, class: "btn btn-warning" %>
   <br>
   <%= render 'form' %>
