+++ app/views/categories/new.html.erb
@@ -1,8 +1,6 @@
 <div class="container">
   <h1>New Category</h1>
-
-  <!-- Render form here -->
-  <%= render 'form' %>
-  <br>
   <%= link_to 'Back', categories_path, class: "btn btn-warning" %>
+  <br>
+  <%= render 'form' %>
 </div>
