+++ app/views/categories/show.html.erb
@@ -54,4 +54,3 @@
     </div>
   <% end %>
 </div>
-<br>
