+++ app/views/orders/edit.html.erb
@@ -1,7 +1,6 @@
 <div class="container">
   <h1>Edit order</h1>
-
-  <%= render 'form' %>
-  <br>
   <%= link_to 'Back', orders_path, class: "btn btn-warning" %>
+  <br>
+  <%= render 'form' %>
 </div>
