+++ app/views/orders/new.html.erb
@@ -1,5 +1,5 @@
-<h1>New order</h1>
-
-<%= render 'form' %>
-
-<%= link_to 'Back', orders_path %>
+<div class="container">
+  <h1>New order</h1>
+  <%= render 'form' %>
+  <%= link_to 'Back', orders_path %>
+</div>
