+++ app/views/pages/home.html.erb
@@ -1,3 +1,14 @@
 <div class="container">
-  <h2>This is the home page</h2>
+  <h1>This is the Etsy Homepage!</h1>
+  <div class="row">
+    <div class="col-md-4">
+      <%= link_to 'Categories', categories_path, class: "btn btn-block btn-primary" %><br>
+    </div>
+    <div class="col-md-4">
+      <%= link_to 'Products', products_path, class: "btn btn-block btn-success" %><br>
+    </div>
+    <div class="col-md-4">
+      <%= link_to 'Orders', orders_path, class: "btn btn-block btn-warning" %><br>
+    </div>
+  </div>
 </div>
