+++ app/views/products/edit.html.erb
@@ -1,9 +1,5 @@
 <div class='container'>
-
   <h1>Edit Product</h1>
-
-  <!-- Render form here -->
   <%= render 'form' %>
   <%= link_to 'Back', products_path, class: "btn btn-warning" %>
-
 </div>
