+++ app/views/products/new.html.erb
@@ -1,8 +1,5 @@
 <div class="container">
   <h1>Add New Product</h1>
-
-  <!-- Render form here -->
+  <%= link_to 'Back', products_path, class: "btn btn-warning" %><br>
   <%= render 'form' %>
-
-  <%= link_to 'Back', products_path, class: "btn btn-warning" %>
 </div>
